+++ pkg/testing/status/fake.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package testing
+package status
 
 import (
 	context "context"
