+++ pkg/testing/status/fake_test.go
@@ -1,4 +1,4 @@
-package testing
+package status
 
 import (
 	context "context"
